// src/types.ts
export type SizeBounds = {
  min: number
  max: number
}
// ...
export type Gusts = {
  /**
   * Should the wind gust?
   * @default true
   */
  active: boolean
  /**
   * The likelihood of the wind changing direction after a gust. 0.0 is never, 1.0 is always.
   * @default 0.25
   */
  changeChance: number

  in: In
  out: Out
}

export type Gravity = {
  /**
   * The angle of gravity, in degrees.
   * @default 90
   */
  angle: number

  /**
   * The strength of gravity.
   * @default 0.7
   */
  strength: number
}

export type Wave = {
  /**
   * The frequency of the wave the snowflakes follow.
   * @default 0.02
   */
  frequency: number

  /**
   * The amplitude of the wave the snowflakes follow.
   * @default 1.0
   */
  amplitude: number
}

export type Wind = {
  /**
   * The angle of the wind, in degrees.
   * @default 0
   */
  angle: number

  /**
   * The strength of the wind.
   * @default 0
   */
  strength: number

  gusts: Gusts
}

export type ConfigLayer = {
  /**
   * A hex string representing the colour of the snowflakes in the foreground.
   * @default '#8d90b7'
   */
  colour: string

  /**
   * A number representing the required density of snowflakes on screen. Note, this is not the actual number of snowflakes.
   * @default 200
   */
  density: number

  /**
   * The size of the snowflakes.
   * @default { min: 1, max: 3 }
   */
  size: SizeBounds

  wave: Wave

  gravity: Gravity

  wind: Wind
}

export interface BaseConfig {
  [key: string]: any

  /**
   * A hex string representing the Background Colour of the canvas.
   * @default '#0d0014'
   */
  // background: string

  /**
   * An array of uniquely configured snowflake layers.
   */
  layers: ConfigLayer[]
}

export interface Config extends BaseConfig {
  /**
   * The element to attach the simulation to.
   * @default '#snowfall'
   */
  attachTo: HTMLElement
}

export type DeepPartial<T> = {
  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]> : T[P]
}

export interface CompleteUserConfig extends BaseConfig {
  /**
   * The element to attach the simulation to, as a string representing the ID.
   */
  attachTo: string

  /**
   * A fully configured array of snowflake layers.
   */
  layers: ConfigLayer[]
}

// UserConfig makes all properties of BaseConfig optional, but `layers` is handled differently
export interface UserConfig
  extends Omit<Partial<CompleteUserConfig>, 'layers'> {
  /**
   * An optional array of uniquely configured snowflake layers,
   * where each layer can be partially configured.
   * @default undefined
   */
  layers?: DeepPartial<ConfigLayer>[]

  /**
   * A string that represents the ID of an element you want to attach snowfall to.
   * @default '#snowfall'
   */
  attachTo?: string
}